lte/alt1250: Fix bug that the recovery process stops
Fixed the bug that the recovery process stops
after a modem reset occurs. There is a situation
where no Ack is returned to usrsock when a modem
reset occurs. In this case, the application will
still be waiting and the recovery process will not work.
Therefore, when a modem reset occurs, use
USRSOCK_IOC_REFLESH to release the waiting state
of the application.