lte/alt1250: Fix bug where an assert would occur
Fixed bug where an assert would occur when data
longer than 1500 bytes was specified in sendto.
If the sendto request exceeds 1500 bytes,
the alt1250 daemon will only read 1500 bytes.
The assert occurs at the timing when the remaining
data is read in the next loop. To avoid this problem,
the remaining data is now discarded using seek.