lte/alt1250: Add reset sequence for LwM2M stub
The updated setting values related to the LwM2M function in the
Alt1250 modem are stored in the NV memory, but the changes in the
setting values will be reflected after the reset. Therefore, when
the initial values are in an unintended state, it is necessary to
reset the system again after entering the setting values.
This commit adds the sequence at reset.